# 2015-2016-api-mobiliteitscentrum/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Mobiliteitscentrum Gent Website API (VOPRO 2015-1016)",
        "description": "API specificatie voor Milestone 3.\n",
        "version": "3.0.0"
    },
    "host": "voproX.ugent.be",
    "schemes": [
        "http"
    ],
    "basePath": "/api",
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/user/{user_id}/point_of_interest/{poi_id}/": {
            "get": {
                "summary": "Returns a specific POI of a user",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "The unique identifier of this user",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "poi_id",
                        "in": "path",
                        "description": "The unique identifier of this POI",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "The access token as a bearer",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The POI",
                        "schema": {
                            "$ref": "#/definitions/PointOfInterest"
                        }
                    },
                    "401": {
                        "description": "The user is not logged in",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "The requested user is not the current user or the current user is not an Administrator",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "User or POI does not exist",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "summary": "Edit a POI",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "The unique identifier of this user",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "poi_id",
                        "in": "path",
                        "description": "The unique identifier of this POI",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "point_of_interest",
                        "in": "body",
                        "description": "The modified POI",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/PointOfInterest"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "The access token as a bearer",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The modified POI",
                        "schema": {
                            "$ref": "#/definitions/PointOfInterest"
                        }
                    },
                    "401": {
                        "description": "The user is not logged in",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "The requested user is not the current user or the current user is not an Administrator",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "User or POI does not exist",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "409": {
                        "description": "The POI already exists",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Deletes a POI",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "The unique identifier of this user",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "poi_id",
                        "in": "path",
                        "description": "The unique identifier of this POI.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "The access token as a bearer",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The POI was deleted"
                    },
                    "401": {
                        "description": "The user is not logged in",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "The requested user is not the current user or the current user is not an Administrator",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "User or POI does not exist",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/event/": {
            "get": {
                "summary": "Returns all events",
                "description": "This endpoint returns all events in JSON format.  TODO: more filters\n",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Only return relevant events for this user",
                        "required": false,
                        "type": "number"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "The access token as a bearer",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An array of events",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Event"
                            }
                        }
                    },
                    "401": {
                        "description": "The user is not logged in",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new event",
                "description": "Creates a new event.\n",
                "parameters": [
                    {
                        "name": "event",
                        "in": "body",
                        "description": "Event input parameters",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Event"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "The access token as a bearer",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "The event is created",
                        "schema": {
                            "$ref": "#/definitions/Event"
                        }
                    },
                    "401": {
                        "description": "The user is not logged in",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "The requested user is not authorized to create events",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Address": {
            "type": "object",
            "properties": {
                "street": {
                    "type": "string"
                },
                "housenumber": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string",
                    "description": "2 letter country code (ex. BE)",
                    "maxLength": 2,
                    "minLength": 2
                },
                "postal_code": {
                    "type": "string"
                },
                "coordinates": {
                    "$ref": "#/definitions/Coordinates"
                }
            },
            "required": [
                "street",
                "housenumber",
                "city",
                "country",
                "postal_code"
            ]
        },
        "Coordinates": {
            "type": "object",
            "properties": {
                "lat": {
                    "type": "number",
                    "format": "double",
                    "description": "Latitude"
                },
                "lon": {
                    "type": "number",
                    "format": "double",
                    "description": "Longitude"
                }
            },
            "required": [
                "lat",
                "lon"
            ]
        },
        "PointOfInterest": {
            "type": "object",
            "properties": {
                "address": {
                    "$ref": "#/definitions/Address"
                },
                "name": {
                    "type": "string",
                    "description": "Human readable name for this address to quickly identify this address (ex. home, work)\n"
                },
                "radius": {
                    "type": "number",
                    "format": "int32",
                    "description": "Radius in meters around home address for which the user wants to recieve notifications"
                },
                "active": {
                    "type": "boolean",
                    "description": "Boolean to indicate if the user wants to receive notifications for events in the radius around this address"
                },
                "notify_for_event_types": {
                    "type": "array",
                    "description": "User has to mark specific EventTypes as relevant for this POI to receive notifications",
                    "items": {
                        "$ref": "#/definitions/EventType"
                    }
                }
            },
            "required": [
                "name",
                "address",
                "radius",
                "active",
                "notify_for_event_types"
            ]
        },
...
        "Event": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "description": "Unique identifier for this Event"
                },
                "links": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Link"
                    }
                },
                "coordinates": {
                    "$ref": "#/definitions/Coordinates"
                },
                "active": {
                    "type": "boolean",
                    "description": "Indicates if this event is active"
                },
                "publication_time": {
                    "type": "string",
                    "format": "date_time",
                    "description": "Time at which this event was created"
                },
                "last_edit_time": {
                    "type": "string",
                    "format": "date_time",
                    "description": "Last time at which this event was updated"
                },
                "description": {
                    "type": "string",
                    "description": "User can enter a description for this event"
                },
                "jams": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "description": "If this event has a traffic jam, this object is filled. Else, this object is null.\n",
                        "properties": {
                            "points": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/Coordinates"
                                }
                            },
                            "speed": {
                                "type": "number",
                                "description": "Average speed of the jam"
                            },
                            "delay": {
                                "type": "number",
                                "description": "Delay of jam, compared to free flow (in seconds)"
                            }
                        }
                    }
                },
                "source": {
                    "type": "object",
                    "description": "The source of this event (ex. Waze, police",
                    "properties": {
                        "name": {
                            "type": "string",
                            "description": "Name of the source which generated this event (ex. Waze, local police)"
                        },
                        "icon_url": {
                            "type": "string",
                            "description": "An URL to the icon of this source"
                        }
                    }
                },
                "type": {
                    "$ref": "#/definitions/EventType"
                },
                "relevant_for_transportation_types": {
                    "type": "array",
                    "description": "An array of transportation types for which this event type is relevant (ex. road block for transportation type car)",
                    "items": {
                        "type": "string",
                        "description": "Tranportation Type"
                    }
                }
            },
            "required": [
                "coordinates",
                "active",
                "description",
                "relevant_for_transportation_types"
            ]
        },
        "EventType": {
            "type": "object",
            "properties": {
                "type": {
                    "type": "string",
                    "description": "Type of event"
                }
            }
        },
...
        "Link": {
            "type": "object",
            "properties": {
                "rel": {
                    "type": "string",
                    "description": "Indicates the relation of the linked object (ex. self)"
                },
                "href": {
                    "type": "string",
                    "description": "The actual link in full URL (ex. https://vopro.ugent.be/api/user/123/)"
                }
            }
        },
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32",
                    "description": "HTTP status code (ex. 400)"
                },
                "message": {
                    "type": "string",
                    "description": "Human readable error (ex. You are not logged in)"
                },
                "fields": {
                    "type": "string",
                    "description": "Missing or incorrect fields for input"
                }
            }
        },
...
    }
}